fix: BigInteger.Compare and arithmetic preserve precision via BigFloat promotion (M5)
Changes BigInteger operations with Float to promote both operands to BigFloat
instead of demoting BigInteger to float64, preventing precision loss for
integers beyond the float64 mantissa limit (2^53 bits).

Fixed methods:
- Compare() with Float and Complex (2 cases)
- Add(), Subtract(), Multiply(), Divide() with Float (4 methods)

Pattern: Mirrors Float.Compare() - always promote to higher-precision type.
Returns BigFloat (inexact) to preserve R7RS §6.2.2 exactness contagion.

Semantic change: BigInteger + Float → BigFloat (was Float).
This is more correct and preserves precision for large integers.

Test coverage:
- New file: values/big_integer_precision_test.go (22 test cases)
- Updated 5 existing test files to expect BigFloat instead of Float
- Added known divergence to lattice model tests

Fixes architectural review finding M5.

+++ values/exactness_contagion_test.go
@@ -57,9 +57,10 @@ func TestExactnessContagionAddition(t *testing.T) {
 		// Integer (exact) + Integer (exact) → Integer (exact)
 		{"Integer 0 + Integer 0", NewInteger(0), NewInteger(0), "Integer", true},
 
-		// BigInteger (exact) + Float (inexact) → Float (inexact)
-		{"BigInteger 0 + Float 0.0", NewBigIntegerFromInt64(0), NewFloat(0.0), "Float", false},
-		{"Float 0.0 + BigInteger 0", NewFloat(0.0), NewBigIntegerFromInt64(0), "Float", false},
+		// BigInteger (exact) + Float (inexact) → BigFloat (inexact)
+		// Changed: precision preservation via BigFloat instead of Float
+		{"BigInteger 0 + Float 0.0", NewBigIntegerFromInt64(0), NewFloat(0.0), "BigFloat", false},
+		{"Float 0.0 + BigInteger 0", NewFloat(0.0), NewBigIntegerFromInt64(0), "BigFloat", false},
 
 		// BigInteger (exact) + BigFloat (inexact) → BigFloat (inexact)
 		{"BigInteger 0 + BigFloat 0.0", NewBigIntegerFromInt64(0), NewBigFloatFromFloat64(0.0), "BigFloat", false},
@@ -127,9 +128,10 @@ func TestExactnessContagionSubtraction(t *testing.T) {
 		// Integer (exact) - Integer (exact) → Integer (exact)
 		{"Integer 0 - Integer 0", NewInteger(0), NewInteger(0), "Integer", true},
 
-		// BigInteger (exact) - Float (inexact) → Float (inexact)
-		{"BigInteger 0 - Float 0.0", NewBigIntegerFromInt64(0), NewFloat(0.0), "Float", false},
-		{"Float 0.0 - BigInteger 0", NewFloat(0.0), NewBigIntegerFromInt64(0), "Float", false},
+		// BigInteger (exact) - Float (inexact) → BigFloat (inexact)
+		// Changed: precision preservation via BigFloat instead of Float
+		{"BigInteger 0 - Float 0.0", NewBigIntegerFromInt64(0), NewFloat(0.0), "BigFloat", false},
+		{"Float 0.0 - BigInteger 0", NewFloat(0.0), NewBigIntegerFromInt64(0), "BigFloat", false},
 
 		// BigInteger (exact) - BigFloat (inexact) → BigFloat (inexact)
 		{"BigInteger 0 - BigFloat 0.0", NewBigIntegerFromInt64(0), NewBigFloatFromFloat64(0.0), "BigFloat", false},
